Loot update

+++ MazeRunner/src/mazerunner/Enemy.java
@@ -37,7 +37,7 @@ public class Enemy extends Creature implements VisibleObject{
 	 * @param weapon			The weapon of the Enemy
 	 * @param modelFileLocation The location of the model file
 	 */
-	public Enemy(GL gl,int x, int y, int z, double horAngle, double hitpoints,
+	public Enemy(GL gl,int x, int y, int z, double horAngle, int hitpoints,
 			Weapon weapon, String modelFileLocation) {
 		// Set the initial position and viewing direction of the enemy.
 		super(gl,x*Maze.SQUARE_SIZE + Maze.SQUARE_SIZE/2,
